Allow the use of specific nodePorts
In some setups, external components like a TCP
loadbalancer need to know the nodePorts where the
controller is exposed.

This commit adds the option to define nodePorts
for the default services http, https and stat.

It also adds this option for tcpPorts but this introduces
a breaking change as a tcpPort must now be a map with
name, port, targetPort and optionnaly nodePort.

Signed-off-by: lcavajani <lcavajani@suse.com>