MINOR: kubernetes-ingress: Allow creating metrics service if controller service is not LoadBalancer (#306)
There doesn't seem to be any benefit or reason to *prevent* users from creating
an additional service to monitor metrics, even if the main service is not of
type `LoadBalancer`. Also, same for if the stats port is already exposed.

I'd like to be able to create an additional service for scraping metrics
regardless.

Signed-off-by: Jason Vigil <jason@trychroma.com>